requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
joblib>=1.3.0
Pillow>=9.0.0

